sns-auto-poster/cloud-run/web-ui/requirements.txt
flask==3.*
gunicorn==21.*
requests==2.*
google-cloud-firestore==2.*
